requirements.txt
pysdl2>=0.9.17
customtkinter>=5.2.0
